modified:   database/migrations/2022_09_28_163901_create_products_table.php

+++ database/migrations/2022_09_28_163901_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('category');
             $table->string('title');
             $table->text('description');
-            $table->float('price', 8, 2);
+            $table->decimal('price', 10, 2);
             $table->string('paypal_id')->nullable();
             $table->timestamps();
         });
